# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Attendance System API
  version: 0.1.0
  description: |
    Multi-tenant API scoped by `company_id`.
    When authentication is enabled, tenant is derived from credentials (for example API key context).
    Request-scoped `company_id` parameters are primarily for development/testing and non-auth deployments.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/simulate/day:
    post:
      tags: [simulation]
      operationId: simulateDay
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Simulation result
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                validation_error:
                  summary: Missing person_id
                  value:
                    error: bad_request
                    code: VALIDATION_ERROR
                    message: Validation failed
                    details:
                      kind: validation
                      error: person_id_required
                      detail: person_id is required
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                internal_error:
                  summary: Unexpected server failure
                  value:
                    error: server_error
                    code: INTERNAL_ERROR
                    message: Server error
  /api/simulate/range:
    post:
      tags: [simulation]
      operationId: simulateRange
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Simulation result
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                validation_error:
                  summary: Missing person_id
                  value:
                    error: bad_request
                    code: VALIDATION_ERROR
                    message: Validation failed
                    details:
                      kind: validation
                      error: person_id_required
                      detail: person_id is required
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                internal_error:
                  summary: Unexpected server failure
                  value:
                    error: server_error
                    code: INTERNAL_ERROR
                    message: Server error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AgentBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details: {}